week11_hw1_updated
