+++ homeworks/week11/hw1/handle_logout.php
