+++ homeworks/week11/hw1/handle_update_comment.php
@@ -2,7 +2,14 @@
   session_start();
   require_once('inc/conn.php');
   require_once('inc/utils.php');
+
   $CSRFToken = $_POST['CSRFToken'];
+
+  if($_POST['CSRFToken'] != $_COOKIE['CSRFToken']) {
+    header('Location: index.php');
+    die();
+  }
+
   if(empty($_POST['content'])) {
     header('Location: update_comment.php?errorCode=1&id='. $_POST['id'] . '&CSRFToken=' . $CSRFToken);
     die();
@@ -13,6 +20,7 @@
   $role = $user['role'];
   $id = $_POST['id'];
   $content = $_POST['content'];
+
   if ($username == 'admin' || $role == '2') {
     $sql = "UPDATE nicolakacha_comments SET content=? WHERE id=?";
     $stmt = $conn->prepare($sql);
