+++ homeworks/week11/hw1/inc/header.php
@@ -3,8 +3,8 @@
   <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Boarddd</title>
-    <link rel="stylesheet" href="inc/style.css">
+    <title>Boarddd v3.2</title>
+    <link rel="stylesheet" href="style/style.css">
   </head>
 
   <body>
