+++ homeworks/week11/hw1/login.php
@@ -29,6 +29,6 @@
       </div>
     </form>
   </main>
-  <script src="inc/script.js"></script>
-  
+
+  <script src="script/script.js"></script>
 <?php include('inc/footer.php'); ?>
